Break misc_non_drivers tests into files
